Feat: spring, django api 연결

+++ server/src/main/java/com/singsong/api/controller/MemberController.java
@@ -61,7 +61,7 @@ public class MemberController {
     }
 
     @PatchMapping("/tag/add")
-    public ResponseEntity<?> memberTagAdd(@RequestParam(value = "tag") int tagId, @ApiIgnore Authentication authentication) {
+    public ResponseEntity<?> memberTagAdd(@RequestParam(value = "tag") Long tagId, @ApiIgnore Authentication authentication) {
         Member member = jwtAuthenticationUtil.jwtTokenAuth(authentication);
         tagService.addMemberTag(member, tagId);
 
@@ -69,7 +69,7 @@ public class MemberController {
     }
 
     @PatchMapping("/tag/delete")
-    public ResponseEntity<?> memberTagDelete(@RequestParam(value = "tag") int tagId, @ApiIgnore Authentication authentication) {
+    public ResponseEntity<?> memberTagDelete(@RequestParam(value = "tag") Long tagId, @ApiIgnore Authentication authentication) {
         Member member = jwtAuthenticationUtil.jwtTokenAuth(authentication);
         tagService.deleteMemberTag(member, tagId);
